logo/requirements copy.txt
numpy==2.2.6
numpy-stl==3.2.0
pillow==11.3.0
pymesh==1.0.2
python-utils==3.9.1
typing_extensions==4.15.0
